combine redundant TaskListItem into TaskList
