--- src/pages/Browse/TaskList/TaskListItem/index.js
@@ -1,1 +1,0 @@
-export { default } from './TaskListItem';
